+++ e2e/config/config.go
@@ -117,9 +117,10 @@ type Contracts struct {
 	Solana Solana `yaml:"solana"`
 }
 
-// Solana contains the addresses of predeployed contracts on the Solana chain
+// Solana contains the addresses of predeployed contracts and accounts on the Solana chain
 type Solana struct {
-	GatewayProgramID string `yaml:"gateway_program_id"`
+	GatewayProgramID string             `yaml:"gateway_program_id"`
+	SPLAddr          DoubleQuotedString `yaml:"spl"`
 }
 
 // EVM contains the addresses of predeployed contracts on the EVM chain
@@ -141,6 +142,7 @@ type ZEVM struct {
 	ERC20ZRC20Addr     DoubleQuotedString `yaml:"erc20_zrc20"`
 	BTCZRC20Addr       DoubleQuotedString `yaml:"btc_zrc20"`
 	SOLZRC20Addr       DoubleQuotedString `yaml:"sol_zrc20"`
+	SPLZRC20Addr       DoubleQuotedString `yaml:"spl_zrc20"`
 	TONZRC20Addr       DoubleQuotedString `yaml:"ton_zrc20"`
 	UniswapFactoryAddr DoubleQuotedString `yaml:"uniswap_factory"`
 	UniswapRouterAddr  DoubleQuotedString `yaml:"uniswap_router"`