+++ e2e/runner/setup_solana.go
@@ -87,6 +87,10 @@ func (r *E2ERunner) SetupSolana(deployerPrivateKey string) {
 
 	err = r.ensureSolanaChainParams()
 	require.NoError(r, err)
+
+	// deploy test spl
+	tokenAccount := r.DeploySPL(&privkey, true)
+	r.SPLAddr = tokenAccount.PublicKey()
 }
 
 func (r *E2ERunner) ensureSolanaChainParams() error {